Use size_t instead of int for the position argument in str_erase(). Mixing these two types is bad because there is no guarantee sizeof(int)==sizeof(size_t). Thanks to JGS for the patch.

+++ libwzd-core/wzd_string.c
@@ -351,7 +351,7 @@ wzd_string_t * str_trim_right(wzd_string_t *str)
  *
  * The rest of the wzd_string_t is shifted down to fill the gap.
  */
-wzd_string_t * str_erase(wzd_string_t * str, size_t pos, int len)
+wzd_string_t * str_erase(wzd_string_t * str, size_t pos, size_t len)
 {
   if (!str || !str->buffer) return NULL;
   if (pos > str->length) return NULL;
